backend/requirements.txt
fastapi==0.68.2
uvicorn==0.17.6
python-jose==3.3.0
python-dotenv==0.19.2